leaderboards implemented
once a user publishes a story, it will be pushed to the leaderboards collection. If they unpublish, it will be removed.

Edits made to the published work will not be reflected on the leaderboards to ensure the work is complete before publishing.

in addition, unpublishing a work will remove all votes and comments. This avoids republishing a winning story to automatically lead in future competitions