updated stories to display amount of ratings and current average rating
Also added more in depth validation for the rating button. If you are the author of the story it will be grayed out, inactive and you will not even be able to select a number out of 10. Text above explains that authors cannot vote on their own stories.

In addition, if the user has already voted the button will be disabled and text will display to let the user know they have already voted.

also added loading indicators and attempted to update the data once a user submits a rating. This did not work.